Restructure cvs (#18)
