+++ pysages/ssages/collective_variables/core.py
@@ -218,163 +218,3 @@ def group_size(obj: range):
 @dispatch
 def group_size(obj: Union[List, Tuple]):
     return sum(group_size(o) for o in obj)
-
-
-# ========== #
-#   Angles   #
-# ========== #
-
-class Angle(ThreePointCV):
-    @property
-    def function(self):
-        return angle
-
-
-def angle(p1, p2, p3):
-    """
-    Returns the angle defined by three points in space
-    (around the one in the middle).
-    """
-    q = p1 - p2
-    r = p3 - p2
-    return np.arctan2(linalg.norm(np.cross(q, r)), np.dot(q, r))
-
-
-class DihedralAngle(FourPointCV):
-    @property
-    def function(self):
-        return dihedral_angle
-
-
-def dihedral_angle(p1, p2, p3, p4):
-    """
-    Returns the dihedral angle defined by four points in space
-    (around the line defined by the two central points).
-    """
-    q = p3 - p2
-    r = np.cross(p2 - p1, q)
-    s = np.cross(q, p4 - p3)
-    return np.arctan2(np.dot(np.cross(r, s), q), np.dot(r, s) * linalg.norm(q))
-
-
-# ===================== #
-#   Shape Descriptors   #
-# ===================== #
-
-def gyration_tensor(positions):
-    n = positions.shape[0]
-    S = np.zeros((3, 3))
-    for r in positions:
-        S += np.outer(r, r)
-    return S / n
-
-
-def weighted_gyration_tensor(positions, weights):
-    n = positions.shape[0]
-    S = np.zeros((3, 3))
-    for i in range(n):
-        w, r = weights[i], positions[i]
-        S += w * np.outer(r, r)
-    return S
-
-
-class RadiusOfGyration(CollectiveVariable):
-    @property
-    def function(self):
-        return radius_of_gyration
-
-
-def radius_of_gyration(positions):
-    n = positions.shape[0]
-    S = np.zeros((3,))
-    # TODO: Replace by `np.sum` and `vmap`
-    for r in positions:
-        S[:] += np.dot(r, r)
-    return S / n
-
-
-def weighted_radius_of_gyration(positions, weights):
-    n = positions.shape[0]
-    R2 = np.zeros((3,))
-    # TODO: Replace by `np.sum` and `vmap`
-    for i in range(n):
-        w, r = weights[i], positions[i]
-        R2 += w * np.dot(r, r)
-    return R2
-
-
-class PrincipalMoment(AxisCV):
-    @property
-    def function(self):
-        return (lambda rs: principal_moments(rs)[self.axis])
-
-
-def principal_moments(positions):
-    return linalg.eigvals(gyration_tensor(positions))
-
-
-class Asphericity(CollectiveVariable):
-    @property
-    def function(self):
-        return asphericity
-
-
-def asphericity(positions):
-    λ1, λ2, λ3 = principal_moments(positions)
-    return λ3 - (λ1 + λ2) / 2
-
-
-class Acylindricity(CollectiveVariable):
-    @property
-    def function(self):
-        return acylindricity
-
-
-def acylindricity(positions):
-    λ1, λ2, _ = principal_moments(positions)
-    return (λ2 - λ1)
-
-
-class ShapeAnysotropy(CollectiveVariable):
-    @property
-    def function(self):
-        return shape_anysotropy
-
-
-def shape_anysotropy(positions):
-    λ1, λ2, λ3 = principal_moments(positions)
-    return (3 * (λ1**2 + λ2**2 + λ3**2) / (λ1 + λ2 + λ3)**2 - 1) / 2
-
-
-# ======================== #
-#   Particle Coordinates   #
-# ======================== #
-
-def barycenter(positions):
-    return np.sum(positions, axis=0) / positions.shape[0]
-
-
-def weighted_barycenter(positions, weights):
-    n = positions.shape[0]
-    R = np.zeros(3)
-    # TODO: Replace by `np.sum` and `vmap`
-    for i in range(n):
-        w, r = weights[i], positions[i]
-        R += w * r
-    return R
-
-
-class Component(AxisCV):
-    @property
-    def function(self):
-        return (lambda rs: barycenter(rs)[self.axis])
-
-
-class Distance(TwoPointCV):
-    @property
-    def function(self):
-        return distance
-
-
-def distance(r1, r2):
-    return linalg.norm(r1 - r2)
